move `migrate-js-config` to `codemods/config/migrate-js-config`
