+++ packages/@tailwindcss-upgrade/src/codemods/css/migrate.ts
@@ -1,8 +1,8 @@
 import postcss from 'postcss'
 import type { Config } from '../../../../tailwindcss/src/compat/plugin-api'
 import type { DesignSystem } from '../../../../tailwindcss/src/design-system'
-import type { JSConfigMigration } from '../../migrate-js-config'
 import { Stylesheet } from '../../stylesheet'
+import type { JSConfigMigration } from '../config/migrate-js-config'
 import { migrateAtApply } from './migrate-at-apply'
 import { migrateAtLayerUtilities } from './migrate-at-layer-utilities'
 import { migrateConfig } from './migrate-config'
