+++ packages/@tailwindcss-upgrade/src/index.ts
@@ -4,6 +4,7 @@ import { globby } from 'globby'
 import fs from 'node:fs/promises'
 import path from 'node:path'
 import postcss from 'postcss'
+import { migrateJsConfig } from './codemods/config/migrate-js-config'
 import { analyze as analyzeStylesheets } from './codemods/css/analyze'
 import { formatNodes } from './codemods/css/format-nodes'
 import { linkConfigs as linkConfigsToStylesheets } from './codemods/css/link'
@@ -13,7 +14,6 @@ import { split as splitStylesheets } from './codemods/css/split'
 import { migrate as migrateTemplate } from './codemods/template/migrate'
 import { prepareConfig } from './codemods/template/prepare-config'
 import { help } from './commands/help'
-import { migrateJsConfig } from './migrate-js-config'
 import { migratePostCSSConfig } from './migrate-postcss'
 import { migratePrettierPlugin } from './migrate-prettier'
 import { Stylesheet } from './stylesheet'
